pyproject.toml
[project]
name = "mcp-server-logseq"
version = "0.0.1"
description = "An MCP server for LogSeq API"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Daniel Zholkovsky", email = "daniel@zholkovsky.com" }]
maintainers = [{ name = "Daniel Zholkovsky", email = "daniel@zholkovsky.com" }]
keywords = ["http", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[project.urls]
homepage = "https://github.com/dailydaniel/logseq-mcp"
repository = "https://github.com/dailydaniel/logseq-mcp"
issues = "https://github.com/dailydaniel/logseq-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-logseq = "mcp_server_logseq.__main__:main"
